Middleware made to check if chatroomalreadyexist
Made a malware that checks if a chatroom already exist with a combination of two users if so a new chatroom between the two of them can not be made.

If middleware is not used in this route then users are able to create an infinite amount of chatrooms whcih is not desired for this project.